+++ examples/simple/index.html
@@ -27,7 +27,7 @@
 <body>
 
     <div id="list">
-        <a href="superstore-arrow.html">Superstore example as an Apache Arrow</a>
+        <a href="superstore.html">Superstore example as an Apache Arrow</a>
         <a href="superstore-hypergrid.html">Editable example</a>
         <a href="superstore-custom-grid.html">Themed example</a>
         <a href="streaming.html">High-frequency streaming random data example</a>
