+++ examples/simple/superstore.html
@@ -17,7 +17,7 @@
     <script src="perspective-viewer.js"></script>
     <script src="perspective-viewer-hypergrid.js"></script>
     <script src="perspective-viewer-datagrid.js"></script>
-    <script src="perspective-viewer-d3fc.js"></script>
+    <script src="perspective-viewer-highcharts.js"></script>
 
     <script src="perspective.js"></script>
