+++ packages/perspective-jupyterlab/package.json
@@ -37,8 +37,7 @@
     "dependencies": {
         "@finos/perspective-viewer": "^0.5.1",
         "@finos/perspective-viewer-d3fc": "^0.5.1",
-        "@finos/perspective-viewer-highcharts": "^0.5.1",
-        "@finos/perspective-viewer-hypergrid": "^0.5.1",
+        "@finos/perspective-viewer-datagrid": "^0.5.1",
         "@jupyter-widgets/base": "^3.0.0",
         "@jupyterlab/application": "^2.0.0",
         "@lumino/application": "^1.7.3",
